pyproject.toml
[project]
name = "axon-core"
version = "1.0.0"
description = "Backend productivo para gestión de infraestructura Axon 88"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-socketio>=5.11.4",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.9.0",
    "httpx>=0.27.2",
    "psutil>=6.1.0",
    "loguru>=0.7.2",
    "python-multipart>=0.0.12",
    "docker>=7.1.0",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
